feat(shell): add startup helper command
Sometimes, I need to re-run the startup scripts of a shell. This
normally happens when a database didn't startup correctly for
some reason. As we don't have an easy way of doing this, I end up
restarting the shell, which implies in closing my editor and losing
important states, like Bazel analysis cache.

Although these cases indicate a bug somewhere in the startup process,
I think it makes sense to give developers a way to re-run it to prevent
blocking them on these types of issues.